refactor(admin): split SystemHealth into SOLID components
